requirements.txt
requests==2.28.1
dotenv-0.21.0
python-telegram-bot==13.14